requirements.txt
ccxt
pandas
numpy
matplotlib
requests
discord.py
